+++ apps/client/package.json
@@ -36,7 +36,7 @@
 		"duckdb-async": "^1.1.1",
 		"install": "^0.13.0",
 		"lucide-svelte": "^0.453.0",
-		"squire": "workspace:*",
+		"products": "workspace:*",
 		"zod": "^3.24.1"
 	}
 }
