+++ apps/client/src/lib/dashboard/types.ts
@@ -1,4 +1,8 @@
-import type { ProductDto, PullRequestDto, SecurityAdvisoryDto } from "squire";
+import type {
+	ProductDto,
+	PullRequestDto,
+	SecurityAdvisoryDto,
+} from "../../../../../packages/products";
 
 export type DashboardSummary = {
 	pullRequests: PullRequestDto[];
