+++ apps/workers/package.json
@@ -5,7 +5,7 @@
 	},
 	"dependencies": {
 		"duckdb-async": "^1.1.1",
-		"squire": "workspace:*",
+		"products": "workspace:*",
 		"squire-github": "workspace:*"
 	}
 }
