+++ packages/products/src/interfaces.ts
