+++ packages/products/src/models.ts
