+++ packages/products/src/repository/insights.test.ts
