+++ packages/products/src/repository/insights.ts
