+++ packages/products/src/service.ts
