+++ packages/products/src/worker.ts
